feat(article): add restaurantId to check ownership on delete article

+++ services/articleService.js
@@ -4,18 +4,18 @@ import Article from "../models/articleModel.js";
 // Service for create new article
 export const createArticle = async (data, restaurantId) => {
     return await Article.findOne({ name: data.name })
-    .then( async article => {
-        if(article) return { status: 404, success: false, message: 'Bad request! Article name already exist!' };
-        else {
-            
-            article = new Article({ ...data,  picture: data.picture || '', restaurantId: restaurantId});
-            await article.save();
-            return { status: 201, success: true, message: "Article created", article: data }
-        }
-    })
-    .catch( error => {
-        return { status: 500, success: false, message: error.message };
-    })
+        .then(async article => {
+            if (article) return { status: 404, success: false, message: 'Bad request! Article name already exist!' };
+            else {
+
+                article = new Article({ ...data, picture: data.picture || '', restaurantId: restaurantId });
+                await article.save();
+                return { status: 201, success: true, message: "Article created", article: data }
+            }
+        })
+        .catch(error => {
+            return { status: 500, success: false, message: error.message };
+        })
 }
 
 // Service for get all articles
@@ -23,18 +23,18 @@ export const listArticles = async (data) => {
     try {
         const minPriceFind = data.minPrice;
         const maxPriceFind = data.maxPrice;
-        if(isNaN(minPriceFind) || isNaN(maxPriceFind) || minPriceFind > maxPriceFind ) return { status: 404, success: true, message: 'Invalid price range!' };
-        if(!data.page) data.page = 1;
-        if(!data.limit) data.limit = 30;
+        if (isNaN(minPriceFind) || isNaN(maxPriceFind) || minPriceFind > maxPriceFind) return { status: 404, success: true, message: 'Invalid price range!' };
+        if (!data.page) data.page = 1;
+        if (!data.limit) data.limit = 30;
         const skipPage = (data.page - 1) * data.limit;
         const articleList = await Article.find()
-        .sort({ [data.sortBy]: 1 })
-        .skip(skipPage)
-        .limit(parseInt(data.limit))
-        .where('price').gte(minPriceFind).lte(maxPriceFind)
-        .exec();
+            .sort({ [data.sortBy]: 1 })
+            .skip(skipPage)
+            .limit(parseInt(data.limit))
+            .where('price').gte(minPriceFind).lte(maxPriceFind)
+            .exec();
         const count = articleList.length;
-        if( articleList && count === 0) return { status: 404, success: true, message: 'There are no article!' };
+        if (articleList && count === 0) return { status: 404, success: true, message: 'There are no article!' };
         else return { status: 200, success: true, page: data.page, limit: data.limit, totalArticles: count, articles: articleList };
     } catch (error) {
         return { status: 500, success: false, message: error };
@@ -44,29 +44,33 @@ export const listArticles = async (data) => {
 // Service for update article
 export const updateArticle = async (id, data) => {
     return await Article.findById(id)
-    .then( async (result) => {
-        if(!result) return { status: 404, success: false, message: 'Article not found!' }
-        else {
-            const newArticle = await Article.findByIdAndUpdate(id, {...data})
-            await newArticle.save();
-            return { status: 200, success: true, message: 'Article updated', article: newArticle }
-        }
-    }).catch((err) => {
-        return { status: 400, success: false, message: err.message }
-    });
+        .then(async (result) => {
+            if (!result) return { status: 404, success: false, message: 'Article not found!' }
+            else {
+                const newArticle = await Article.findByIdAndUpdate(id, { ...data })
+                await newArticle.save();
+                return { status: 200, success: true, message: 'Article updated', article: newArticle }
+            }
+        }).catch((err) => {
+            return { status: 400, success: false, message: err.message }
+        });
 };
 
 // Service for delete article
-export const deleteArticle = async (id) => {
+export const deleteArticle = async (id, userId) => {
     return await Article.findById(id)
-    .then( async article => {
-        if(!article) return { status: 404, succes: false, message: 'Article not found!' };
-        else {
-            await Article.findByIdAndDelete(id);
-            return { status: 200, succes: true, message: 'Article is deleted' }
-        }
-    })
-    .catch(err => {
-        return { status: 400, succes: false, message: err }
-    });
+        .then(async article => {
+            if (!article) return { status: 404, succes: false, message: 'Article not found!' };
+            else {
+                const restaurant = article.restaurantId;
+                if (restaurant != userId) return { status: 401, succes: false, message: 'Unauthorized! Invalid token' };
+                else    {
+                    await Article.findByIdAndDelete(id);
+                    return { status: 200, succes: true, message: 'Article is deleted' }
+                };
+            }
+        })
+        .catch(err => {
+            return { status: 400, succes: false, message: err }
+        });
 };
